add conditions for schema migrations

+++ database/migrations/2024_06_24_120624_create_local_market_order_has_units_table.php
@@ -13,16 +13,18 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('local_market_order_has_units', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('local_market_order_id');
-            $table->foreign('local_market_order_id')->references('id')->on('local_market_orders')->onDelete('cascade');
-            $table->unsignedBigInteger('inventory_id');
-            $table->foreign('inventory_id')->references('id')->on('local_market_inventories')->onDelete('cascade');
-            $table->unsignedBigInteger('unit_id');
-            $table->foreign('unit_id')->references('id')->on('local_market_inventory_units')->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (! Schema::hasTable('local_market_order_has_units')) {
+            Schema::create('local_market_order_has_units', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('local_market_order_id');
+                $table->foreign('local_market_order_id')->references('id')->on('local_market_orders')->onDelete('cascade');
+                $table->unsignedBigInteger('inventory_id');
+                $table->foreign('inventory_id')->references('id')->on('local_market_inventories')->onDelete('cascade');
+                $table->unsignedBigInteger('unit_id');
+                $table->foreign('unit_id')->references('id')->on('local_market_inventory_units')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
